+++ src/apis/object_store_api.rs
@@ -17,25 +17,10 @@ pub struct CreateStoreParams {
     pub store: Option<crate::models::Store>
 }
 
-/// struct for passing parameters to the method [`delete_key_from_store`]
-#[derive(Clone, Debug, Default)]
-pub struct DeleteKeyFromStoreParams {
-    pub store_id: String,
-    pub key_name: String
-}
-
 /// struct for passing parameters to the method [`delete_store`]
 #[derive(Clone, Debug, Default)]
 pub struct DeleteStoreParams {
     pub store_id: String
-}
-
-/// struct for passing parameters to the method [`get_keys`]
-#[derive(Clone, Debug, Default)]
-pub struct GetKeysParams {
-    pub store_id: String,
-    pub cursor: Option<String>,
-    pub limit: Option<i32>
 }
 
 /// struct for passing parameters to the method [`get_store`]
@@ -51,21 +36,6 @@ pub struct GetStoresParams {
     pub limit: Option<i32>
 }
 
-/// struct for passing parameters to the method [`get_value_for_key`]
-#[derive(Clone, Debug, Default)]
-pub struct GetValueForKeyParams {
-    pub store_id: String,
-    pub key_name: String
-}
-
-/// struct for passing parameters to the method [`set_value_for_key`]
-#[derive(Clone, Debug, Default)]
-pub struct SetValueForKeyParams {
-    pub store_id: String,
-    pub key_name: String,
-    pub body: Option<std::path::PathBuf>
-}
-
 
 /// struct for typed errors of method [`create_store`]
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -74,24 +44,10 @@ pub enum CreateStoreError {
     UnknownValue(serde_json::Value),
 }
 
-/// struct for typed errors of method [`delete_key_from_store`]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum DeleteKeyFromStoreError {
-    UnknownValue(serde_json::Value),
-}
-
 /// struct for typed errors of method [`delete_store`]
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum DeleteStoreError {
-    UnknownValue(serde_json::Value),
-}
-
-/// struct for typed errors of method [`get_keys`]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum GetKeysError {
     UnknownValue(serde_json::Value),
 }
 
@@ -106,20 +62,6 @@ pub enum GetStoreError {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum GetStoresError {
-    UnknownValue(serde_json::Value),
-}
-
-/// struct for typed errors of method [`get_value_for_key`]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum GetValueForKeyError {
-    UnknownValue(serde_json::Value),
-}
-
-/// struct for typed errors of method [`set_value_for_key`]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum SetValueForKeyError {
     UnknownValue(serde_json::Value),
 }
 
@@ -177,60 +119,7 @@ pub async fn create_store(configuration: &mut configuration::Configuration, para
     }
 }
 
-/// Delete a key from a customer store.
-pub async fn delete_key_from_store(configuration: &mut configuration::Configuration, params: DeleteKeyFromStoreParams) -> Result<(), Error<DeleteKeyFromStoreError>> {
-    let local_var_configuration = configuration;
-
-    // unbox the parameters
-    let store_id = params.store_id;
-    let key_name = params.key_name;
-
-
-    let local_var_client = &local_var_configuration.client;
-
-    let local_var_uri_str = format!("{}/resources/stores/object/{store_id}/keys/{key_name}", local_var_configuration.base_path, store_id=crate::apis::urlencode(store_id), key_name=crate::apis::urlencode(key_name));
-    let mut local_var_req_builder = local_var_client.request(reqwest::Method::DELETE, local_var_uri_str.as_str());
-
-    if let Some(ref local_var_user_agent) = local_var_configuration.user_agent {
-        local_var_req_builder = local_var_req_builder.header(reqwest::header::USER_AGENT, local_var_user_agent.clone());
-    }
-    if let Some(ref local_var_apikey) = local_var_configuration.api_key {
-        let local_var_key = local_var_apikey.key.clone();
-        let local_var_value = match local_var_apikey.prefix {
-            Some(ref local_var_prefix) => format!("{} {}", local_var_prefix, local_var_key),
-            None => local_var_key,
-        };
-        local_var_req_builder = local_var_req_builder.header("Fastly-Key", local_var_value);
-    };
-
-    let local_var_req = local_var_req_builder.build()?;
-    let local_var_resp = local_var_client.execute(local_var_req).await?;
-
-    if "DELETE" != "GET" && "DELETE" != "HEAD" {
-      let headers = local_var_resp.headers();
-      local_var_configuration.rate_limit_remaining = match headers.get("Fastly-RateLimit-Remaining") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => configuration::DEFAULT_RATELIMIT,
-      };
-      local_var_configuration.rate_limit_reset = match headers.get("Fastly-RateLimit-Reset") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => 0,
-      };
-    }
-
-    let local_var_status = local_var_resp.status();
-    let local_var_content = local_var_resp.text().await?;
-
-    if !local_var_status.is_client_error() && !local_var_status.is_server_error() {
-        Ok(())
-    } else {
-        let local_var_entity: Option<DeleteKeyFromStoreError> = serde_json::from_str(&local_var_content).ok();
-        let local_var_error = ResponseContent { status: local_var_status, content: local_var_content, entity: local_var_entity };
-        Err(Error::ResponseError(local_var_error))
-    }
-}
-
-/// An object store must be empty before it can be deleted.  Deleting an object store that still contains keys will result in a 409 Conflict.
+/// An object store must be empty before it can be deleted.  Deleting an object store that still contains keys will result in a `409` (Conflict).
 pub async fn delete_store(configuration: &mut configuration::Configuration, params: DeleteStoreParams) -> Result<(), Error<DeleteStoreError>> {
     let local_var_configuration = configuration;
 
@@ -277,66 +166,6 @@ pub async fn delete_store(configuration: &mut configuration::Configuration, para
         Ok(())
     } else {
         let local_var_entity: Option<DeleteStoreError> = serde_json::from_str(&local_var_content).ok();
-        let local_var_error = ResponseContent { status: local_var_status, content: local_var_content, entity: local_var_entity };
-        Err(Error::ResponseError(local_var_error))
-    }
-}
-
-/// List all keys within an object store.
-pub async fn get_keys(configuration: &mut configuration::Configuration, params: GetKeysParams) -> Result<crate::models::KeyResponse, Error<GetKeysError>> {
-    let local_var_configuration = configuration;
-
-    // unbox the parameters
-    let store_id = params.store_id;
-    let cursor = params.cursor;
-    let limit = params.limit;
-
-
-    let local_var_client = &local_var_configuration.client;
-
-    let local_var_uri_str = format!("{}/resources/stores/object/{store_id}/keys", local_var_configuration.base_path, store_id=crate::apis::urlencode(store_id));
-    let mut local_var_req_builder = local_var_client.request(reqwest::Method::GET, local_var_uri_str.as_str());
-
-    if let Some(ref local_var_str) = cursor {
-        local_var_req_builder = local_var_req_builder.query(&[("cursor", &local_var_str.to_string())]);
-    }
-    if let Some(ref local_var_str) = limit {
-        local_var_req_builder = local_var_req_builder.query(&[("limit", &local_var_str.to_string())]);
-    }
-    if let Some(ref local_var_user_agent) = local_var_configuration.user_agent {
-        local_var_req_builder = local_var_req_builder.header(reqwest::header::USER_AGENT, local_var_user_agent.clone());
-    }
-    if let Some(ref local_var_apikey) = local_var_configuration.api_key {
-        let local_var_key = local_var_apikey.key.clone();
-        let local_var_value = match local_var_apikey.prefix {
-            Some(ref local_var_prefix) => format!("{} {}", local_var_prefix, local_var_key),
-            None => local_var_key,
-        };
-        local_var_req_builder = local_var_req_builder.header("Fastly-Key", local_var_value);
-    };
-
-    let local_var_req = local_var_req_builder.build()?;
-    let local_var_resp = local_var_client.execute(local_var_req).await?;
-
-    if "GET" != "GET" && "GET" != "HEAD" {
-      let headers = local_var_resp.headers();
-      local_var_configuration.rate_limit_remaining = match headers.get("Fastly-RateLimit-Remaining") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => configuration::DEFAULT_RATELIMIT,
-      };
-      local_var_configuration.rate_limit_reset = match headers.get("Fastly-RateLimit-Reset") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => 0,
-      };
-    }
-
-    let local_var_status = local_var_resp.status();
-    let local_var_content = local_var_resp.text().await?;
-
-    if !local_var_status.is_client_error() && !local_var_status.is_server_error() {
-        serde_json::from_str(&local_var_content).map_err(Error::from)
-    } else {
-        let local_var_entity: Option<GetKeysError> = serde_json::from_str(&local_var_content).ok();
         let local_var_error = ResponseContent { status: local_var_status, content: local_var_content, entity: local_var_entity };
         Err(Error::ResponseError(local_var_error))
     }
@@ -395,7 +224,7 @@ pub async fn get_store(configuration: &mut configuration::Configuration, params:
 }
 
 /// Get all stores for a given customer.
-pub async fn get_stores(configuration: &mut configuration::Configuration, params: GetStoresParams) -> Result<crate::models::GetStoresResponse, Error<GetStoresError>> {
+pub async fn get_stores(configuration: &mut configuration::Configuration, params: GetStoresParams) -> Result<crate::models::InlineResponse2002, Error<GetStoresError>> {
     let local_var_configuration = configuration;
 
     // unbox the parameters
@@ -448,114 +277,6 @@ pub async fn get_stores(configuration: &mut configuration::Configuration, params
         serde_json::from_str(&local_var_content).map_err(Error::from)
     } else {
         let local_var_entity: Option<GetStoresError> = serde_json::from_str(&local_var_content).ok();
-        let local_var_error = ResponseContent { status: local_var_status, content: local_var_content, entity: local_var_entity };
-        Err(Error::ResponseError(local_var_error))
-    }
-}
-
-/// Get the value associated with a key.
-pub async fn get_value_for_key(configuration: &mut configuration::Configuration, params: GetValueForKeyParams) -> Result<std::path::PathBuf, Error<GetValueForKeyError>> {
-    let local_var_configuration = configuration;
-
-    // unbox the parameters
-    let store_id = params.store_id;
-    let key_name = params.key_name;
-
-
-    let local_var_client = &local_var_configuration.client;
-
-    let local_var_uri_str = format!("{}/resources/stores/object/{store_id}/keys/{key_name}", local_var_configuration.base_path, store_id=crate::apis::urlencode(store_id), key_name=crate::apis::urlencode(key_name));
-    let mut local_var_req_builder = local_var_client.request(reqwest::Method::GET, local_var_uri_str.as_str());
-
-    if let Some(ref local_var_user_agent) = local_var_configuration.user_agent {
-        local_var_req_builder = local_var_req_builder.header(reqwest::header::USER_AGENT, local_var_user_agent.clone());
-    }
-    if let Some(ref local_var_apikey) = local_var_configuration.api_key {
-        let local_var_key = local_var_apikey.key.clone();
-        let local_var_value = match local_var_apikey.prefix {
-            Some(ref local_var_prefix) => format!("{} {}", local_var_prefix, local_var_key),
-            None => local_var_key,
-        };
-        local_var_req_builder = local_var_req_builder.header("Fastly-Key", local_var_value);
-    };
-
-    let local_var_req = local_var_req_builder.build()?;
-    let local_var_resp = local_var_client.execute(local_var_req).await?;
-
-    if "GET" != "GET" && "GET" != "HEAD" {
-      let headers = local_var_resp.headers();
-      local_var_configuration.rate_limit_remaining = match headers.get("Fastly-RateLimit-Remaining") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => configuration::DEFAULT_RATELIMIT,
-      };
-      local_var_configuration.rate_limit_reset = match headers.get("Fastly-RateLimit-Reset") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => 0,
-      };
-    }
-
-    let local_var_status = local_var_resp.status();
-    let local_var_content = local_var_resp.text().await?;
-
-    if !local_var_status.is_client_error() && !local_var_status.is_server_error() {
-        serde_json::from_str(&local_var_content).map_err(Error::from)
-    } else {
-        let local_var_entity: Option<GetValueForKeyError> = serde_json::from_str(&local_var_content).ok();
-        let local_var_error = ResponseContent { status: local_var_status, content: local_var_content, entity: local_var_entity };
-        Err(Error::ResponseError(local_var_error))
-    }
-}
-
-/// Insert a new key-value pair into an object store.
-pub async fn set_value_for_key(configuration: &mut configuration::Configuration, params: SetValueForKeyParams) -> Result<std::path::PathBuf, Error<SetValueForKeyError>> {
-    let local_var_configuration = configuration;
-
-    // unbox the parameters
-    let store_id = params.store_id;
-    let key_name = params.key_name;
-    let body = params.body;
-
-
-    let local_var_client = &local_var_configuration.client;
-
-    let local_var_uri_str = format!("{}/resources/stores/object/{store_id}/keys/{key_name}", local_var_configuration.base_path, store_id=crate::apis::urlencode(store_id), key_name=crate::apis::urlencode(key_name));
-    let mut local_var_req_builder = local_var_client.request(reqwest::Method::PUT, local_var_uri_str.as_str());
-
-    if let Some(ref local_var_user_agent) = local_var_configuration.user_agent {
-        local_var_req_builder = local_var_req_builder.header(reqwest::header::USER_AGENT, local_var_user_agent.clone());
-    }
-    if let Some(ref local_var_apikey) = local_var_configuration.api_key {
-        let local_var_key = local_var_apikey.key.clone();
-        let local_var_value = match local_var_apikey.prefix {
-            Some(ref local_var_prefix) => format!("{} {}", local_var_prefix, local_var_key),
-            None => local_var_key,
-        };
-        local_var_req_builder = local_var_req_builder.header("Fastly-Key", local_var_value);
-    };
-    local_var_req_builder = local_var_req_builder.json(&body);
-
-    let local_var_req = local_var_req_builder.build()?;
-    let local_var_resp = local_var_client.execute(local_var_req).await?;
-
-    if "PUT" != "GET" && "PUT" != "HEAD" {
-      let headers = local_var_resp.headers();
-      local_var_configuration.rate_limit_remaining = match headers.get("Fastly-RateLimit-Remaining") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => configuration::DEFAULT_RATELIMIT,
-      };
-      local_var_configuration.rate_limit_reset = match headers.get("Fastly-RateLimit-Reset") {
-          Some(v) => v.to_str().unwrap().parse().unwrap(),
-          None => 0,
-      };
-    }
-
-    let local_var_status = local_var_resp.status();
-    let local_var_content = local_var_resp.text().await?;
-
-    if !local_var_status.is_client_error() && !local_var_status.is_server_error() {
-        serde_json::from_str(&local_var_content).map_err(Error::from)
-    } else {
-        let local_var_entity: Option<SetValueForKeyError> = serde_json::from_str(&local_var_content).ok();
         let local_var_error = ResponseContent { status: local_var_status, content: local_var_content, entity: local_var_entity };
         Err(Error::ResponseError(local_var_error))
     }