+++ src/models/tls_subscription_response_attributes_all_of.rs
@@ -34,6 +34,8 @@ pub enum State {
     Issued,
     #[serde(rename = "renewing")]
     Renewing,
+    #[serde(rename = "failed")]
+    Failed,
 }
 
 impl Default for State {